Fixed Kill Death Ratio not showing to 2 d.p

+++ app/src/main/java/com/itachi1706/hypixelstatistics/AsyncAPI/Players/GetPlayerByNameExpanded.java
@@ -828,7 +828,7 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int waDeaths = obj.get("deaths").getAsInt();
             if (waDeaths == 0)
                 waDeaths = 1;  //Done to prevent Divide by Zero Exception
-            double waKDA = waKills / waDeaths;
+            double waKDA = (double) waKills / waDeaths;
             waKDA = (double) Math.round(waKDA * 100) / 100;
             descArray.add(new ResultDescription("K/D Ratio", waKDA + ""));
         }
@@ -875,8 +875,8 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int w3Deaths = obj.get("deaths").getAsInt();
             if (w3Deaths == 0)
                 w3Deaths = 1;  //Done to prevent Divide by Zero Exception
-            double w3KDA = w3Kills / w3Deaths;
-            w3KDA = (double) Math.round(w3KDA * 100) / 100;
+            double w3KDA = (double) w3Kills / w3Deaths;
+            w3KDA = (double) Math.round(w3KDA * 100.00) / 100.00;
             descArray.add(new ResultDescription("K/D Ratio", w3KDA + ""));
         }
         if (obj.has("finalDeaths"))
@@ -1396,7 +1396,7 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int pbDeaths = obj.get("deaths").getAsInt();
             if (pbDeaths == 0)
                 pbDeaths = 1;  //Done to prevent Divide by Zero Exception
-            double pbKDA = pbKills / pbDeaths;
+            double pbKDA = (double) pbKills / pbDeaths;
             pbKDA = (double) Math.round(pbKDA * 100) / 100;
             descArray.add(new ResultDescription("K/D Ratio", pbKDA + ""));
         }
@@ -1445,7 +1445,7 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int mcgoDeaths = obj.get("deaths").getAsInt();
             if (mcgoDeaths == 0)
                 mcgoDeaths = 1;  //Done to prevent Divide by Zero Exception
-            double mcgoKDA = mcgoKills / mcgoDeaths;
+            double mcgoKDA = (double) mcgoKills / mcgoDeaths;
             mcgoKDA = (double) Math.round(mcgoKDA * 100) / 100;
             descArray.add(new ResultDescription("K/D Ratio", mcgoKDA + ""));
         }
@@ -1508,7 +1508,7 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int uhcDeaths = obj.get("deaths").getAsInt();
             if (uhcDeaths == 0)
                 uhcDeaths = 1;  //Done to prevent Divide by Zero Exception
-            double uhcKDA = uhcKills / uhcDeaths;
+            double uhcKDA = (double) uhcKills / uhcDeaths;
             uhcKDA = (double) Math.round(uhcKDA * 100) / 100;
             descArray.add(new ResultDescription("K/D Ratio", uhcKDA + ""));
         }
@@ -1569,7 +1569,7 @@ public class GetPlayerByNameExpanded extends AsyncTask<String,Void,String> {
             int warDeaths = obj.get("deaths").getAsInt();
             if (warDeaths == 0)
                 warDeaths = 1;  //Done to prevent Divide by Zero Exception
-            double warKDA = (warKills + warAssists) / warDeaths;
+            double warKDA = (double) (warKills + warAssists) / warDeaths;
             warKDA = (double) Math.round(warKDA * 100) / 100;
             descArray.add(new ResultDescription("K/D/A Ratio", warKDA + ""));
         }
